LPS-140066 Refactor join in a single module all integration test
