+++ modules/apps/template/template-test/src/testIntegration/java/com/liferay/template/web/internal/internal/portlet/action/test/DeleteTemplateEntryMVCActionCommandTest.java
@@ -12,17 +12,15 @@
  * details.
  */
 
-package com.liferay.template.web.internal.portlet.action.test;
+package com.liferay.template.web.internal.internal.portlet.action.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.dynamic.data.mapping.model.DDMTemplate;
 import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
 import com.liferay.info.item.InfoItemServiceTracker;
-import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
 import com.liferay.portal.kernel.service.CompanyLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
@@ -33,22 +31,13 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 import com.liferay.template.model.TemplateEntry;
 import com.liferay.template.service.TemplateEntryLocalService;
-
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Stream;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -65,7 +54,7 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class CopyTemplateEntryMVCActionCommandTest {
+public class DeleteTemplateEntryMVCActionCommandTest {
 
 	@ClassRule
 	@Rule
@@ -92,25 +81,22 @@ public class CopyTemplateEntryMVCActionCommandTest {
 	}
 
 	@Test
-	public void testCopyTemplateEntry() throws Exception {
+	public void testBulkDeleteTemplateEntry() throws Exception {
 		MockLiferayPortletActionRequest mockLiferayPortletActionRequest =
-			_getMockLiferayPortletActionRequest();
+			new MockLiferayPortletActionRequest();
 
-		String name = RandomTestUtil.randomString();
-
-		String description = RandomTestUtil.randomString();
-
-		String languageId = LocaleUtil.toLanguageId(
-			_portal.getSiteDefaultLocale(_group.getGroupId()));
+		TemplateEntry templateEntry = TemplateTestUtil.addAnyTemplateEntry(
+			_infoItemServiceTracker, _serviceContext);
 
 		mockLiferayPortletActionRequest.addParameter(
-			"name_" + languageId, name);
-		mockLiferayPortletActionRequest.addParameter(
-			"description_" + languageId, description);
+			"rowIds",
+			new String[] {
+				String.valueOf(_templateEntry.getTemplateEntryId()),
+				String.valueOf(templateEntry.getTemplateEntryId())
+			});
 
-		mockLiferayPortletActionRequest.addParameter(
-			"templateEntryId",
-			String.valueOf(_templateEntry.getTemplateEntryId()));
+		_assertTemplateExists(_templateEntry);
+		_assertTemplateExists(templateEntry);
 
 		ReflectionTestUtil.invoke(
 			_mvcActionCommand, "doTransactionalCommand",
@@ -118,66 +104,48 @@ public class CopyTemplateEntryMVCActionCommandTest {
 			mockLiferayPortletActionRequest,
 			new MockLiferayPortletActionResponse());
 
-		List<TemplateEntry> templateEntries =
-			_templateEntryLocalService.getTemplateEntries(
-				_group.getGroupId(), _templateEntry.getInfoItemClassName(),
-				_templateEntry.getInfoItemFormVariationKey(), QueryUtil.ALL_POS,
-				QueryUtil.ALL_POS, null);
-
-		Assert.assertTrue(templateEntries.size() > 1);
-
-		DDMTemplate originalDDMTemplate =
-			_ddmTemplateLocalService.getDDMTemplate(
-				_templateEntry.getDDMTemplateId());
-
-		Stream<TemplateEntry> templateEntriesStream = templateEntries.stream();
-
-		Assert.assertTrue(
-			templateEntriesStream.map(
-				templateEntry -> _ddmTemplateLocalService.fetchDDMTemplate(
-					templateEntry.getDDMTemplateId())
-			).filter(
-				Objects::nonNull
-			).filter(
-				ddmTemplate -> Objects.equals(
-					name, ddmTemplate.getName(languageId))
-			).filter(
-				ddmTemplate -> Objects.equals(
-					description, ddmTemplate.getDescription(languageId))
-			).filter(
-				ddmTemplate -> Objects.equals(
-					originalDDMTemplate.getScript(), ddmTemplate.getScript())
-			).findAny(
-			).isPresent());
+		_assertTemplateNotExists(_templateEntry);
+		_assertTemplateNotExists(templateEntry);
 	}
 
-	private MockLiferayPortletActionRequest
-			_getMockLiferayPortletActionRequest()
-		throws Exception {
-
+	@Test
+	public void testDeleteTemplateEntry() throws Exception {
 		MockLiferayPortletActionRequest mockLiferayPortletActionRequest =
 			new MockLiferayPortletActionRequest();
 
-		mockLiferayPortletActionRequest.setAttribute(
-			WebKeys.THEME_DISPLAY, _getThemeDisplay());
-
 		mockLiferayPortletActionRequest.addParameter(
-			"groupId", String.valueOf(_group.getGroupId()));
+			"templateEntryId",
+			String.valueOf(_templateEntry.getTemplateEntryId()));
 
-		return mockLiferayPortletActionRequest;
+		_assertTemplateExists(_templateEntry);
+
+		ReflectionTestUtil.invoke(
+			_mvcActionCommand, "doTransactionalCommand",
+			new Class<?>[] {ActionRequest.class, ActionResponse.class},
+			mockLiferayPortletActionRequest,
+			new MockLiferayPortletActionResponse());
+
+		_assertTemplateNotExists(_templateEntry);
 	}
 
-	private ThemeDisplay _getThemeDisplay() throws Exception {
-		ThemeDisplay themeDisplay = new ThemeDisplay();
+	private void _assertTemplateExists(TemplateEntry templateEntry) {
+		Assert.assertNotNull(
+			_templateEntryLocalService.fetchTemplateEntry(
+				templateEntry.getTemplateEntryId()));
 
-		themeDisplay.setCompany(_company);
-		themeDisplay.setPermissionChecker(
-			PermissionThreadLocal.getPermissionChecker());
-		themeDisplay.setScopeGroupId(_group.getGroupId());
-		themeDisplay.setSiteGroupId(_group.getGroupId());
-		themeDisplay.setUser(TestPropsValues.getUser());
+		Assert.assertNotNull(
+			_ddmTemplateLocalService.fetchDDMTemplate(
+				templateEntry.getDDMTemplateId()));
+	}
 
-		return themeDisplay;
+	private void _assertTemplateNotExists(TemplateEntry templateEntry) {
+		Assert.assertNull(
+			_templateEntryLocalService.fetchTemplateEntry(
+				templateEntry.getTemplateEntryId()));
+
+		Assert.assertNull(
+			_ddmTemplateLocalService.fetchDDMTemplate(
+				templateEntry.getDDMTemplateId()));
 	}
 
 	private Company _company;
@@ -197,11 +165,8 @@ public class CopyTemplateEntryMVCActionCommandTest {
 	@Inject
 	private InfoItemServiceTracker _infoItemServiceTracker;
 
-	@Inject(filter = "mvc.command.name=/template/copy_template_entry")
+	@Inject(filter = "mvc.command.name=/template/delete_template_entry")
 	private MVCActionCommand _mvcActionCommand;
-
-	@Inject
-	private Portal _portal;
 
 	private ServiceContext _serviceContext;
